+++ Changelog.md
@@ -5,7 +5,28 @@ All notable changes to this Julia package will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.0] - 18/02/2023
+## [0.1.1] – 16/09/2023
+
+### Added
+
+* Rayleigh Quotient functions added
+* an example illustrating Euclidean gradient/HEssian conversion
+* Add Literature with DocumenterCitations
+
+## [0.1.2] – 13/06/2023
+
+### Added
+
+* Update examples to use Quarto
+* Add DC examples
+
+## [0.1.1] – 01/03/2023
+
+### Added
+
+* Rosenbrock function and examples
+
+## [0.1.0] – 18/02/2023
 
 ### Added